# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Project API",
    "description": "筋トレ・競馬収支・予算管理 API",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/google/callback": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Google Callback",
        "operationId": "google_callback_auth_google_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleCallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workout/weight": {
      "post": {
        "tags": [
          "workout"
        ],
        "summary": "Upsert Weight",
        "operationId": "upsert_weight_workout_weight_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightRecordCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workout"
        ],
        "summary": "List Weight",
        "operationId": "list_weight_workout_weight_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightRecordListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workout/weight/{record_date}": {
      "delete": {
        "tags": [
          "workout"
        ],
        "summary": "Delete Weight",
        "operationId": "delete_weight_workout_weight__record_date__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Date"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workout/exercise": {
      "get": {
        "tags": [
          "workout"
        ],
        "summary": "List Exercises",
        "operationId": "list_exercises_workout_exercise_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "workout"
        ],
        "summary": "Create Exercise",
        "operationId": "create_exercise_workout_exercise_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/workout/exercise/{exercise_id}": {
      "delete": {
        "tags": [
          "workout"
        ],
        "summary": "Delete Exercise",
        "operationId": "delete_exercise_workout_exercise__exercise_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workout/calorie": {
      "post": {
        "tags": ["workout"],
        "summary": "Upsert Calorie",
        "operationId": "upsert_calorie_workout_calorie_post",
        "security": [{"HTTPBearer": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CalorieLogCreate"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CalorieLogResponse"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      },
      "get": {
        "tags": ["workout"],
        "summary": "List Calorie",
        "operationId": "list_calorie_workout_calorie_get",
        "security": [{"HTTPBearer": []}],
        "parameters": [
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From Date"}
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "To Date"}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CalorieLogListResponse"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/workout/calorie/{record_date}": {
      "delete": {
        "tags": ["workout"],
        "summary": "Delete Calorie",
        "operationId": "delete_calorie_workout_calorie__record_date__delete",
        "security": [{"HTTPBearer": []}],
        "parameters": [
          {
            "name": "record_date",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "title": "Record Date"}
          }
        ],
        "responses": {
          "204": {"description": "Successful Response"},
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/workout/log": {
      "post": {
        "tags": ["workout"],
        "summary": "Create Log",
        "operationId": "create_log_workout_log_post",
        "security": [{"HTTPBearer": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/WorkoutLogCreate"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/WorkoutLogResponse"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      },
      "get": {
        "tags": ["workout"],
        "summary": "List Logs",
        "operationId": "list_logs_workout_log_get",
        "security": [{"HTTPBearer": []}],
        "parameters": [
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From Date"}
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "To Date"}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/WorkoutLogListResponse"}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/workout/log/{log_id}": {
      "delete": {
        "tags": ["workout"],
        "summary": "Delete Log",
        "operationId": "delete_log_workout_log__log_id__delete",
        "security": [{"HTTPBearer": []}],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "title": "Log Id"}
          }
        ],
        "responses": {
          "204": {"description": "Successful Response"},
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ExerciseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "muscle_group": {
            "$ref": "#/components/schemas/MuscleGroup"
          },
          "calories_per_rep_per_kg": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories Per Rep Per Kg"
          }
        },
        "type": "object",
        "required": [
          "name",
          "muscle_group"
        ],
        "title": "ExerciseCreate"
      },
      "ExerciseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExerciseResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExerciseListResponse"
      },
      "ExerciseResponse": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "title": "Exercise Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "muscle_group": {
            "$ref": "#/components/schemas/MuscleGroup"
          },
          "calories_per_rep_per_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories Per Rep Per Kg"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "name",
          "muscle_group",
          "calories_per_rep_per_kg",
          "created_at",
          "updated_at"
        ],
        "title": "ExerciseResponse"
      },
      "GoogleCallbackRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "code_verifier": {
            "type": "string",
            "title": "Code Verifier"
          }
        },
        "type": "object",
        "required": [
          "code",
          "code_verifier"
        ],
        "title": "GoogleCallbackRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MuscleGroup": {
        "type": "string",
        "enum": [
          "chest",
          "back",
          "shoulder",
          "arm",
          "abdomen",
          "leg",
          "other"
        ],
        "title": "MuscleGroup"
      },
      "TokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "TokenResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "CalorieLogCreate": {
        "properties": {
          "record_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Record Date"
          },
          "calories": {
            "type": "integer",
            "minimum": 0,
            "maximum": 99999,
            "title": "Calories"
          }
        },
        "type": "object",
        "required": ["record_date", "calories"],
        "title": "CalorieLogCreate"
      },
      "CalorieLogResponse": {
        "properties": {
          "record_date": {"type": "string", "title": "Record Date"},
          "calories": {"type": "integer", "title": "Calories"},
          "created_at": {"type": "string", "title": "Created At"},
          "updated_at": {"type": "string", "title": "Updated At"}
        },
        "type": "object",
        "required": ["record_date", "calories", "created_at", "updated_at"],
        "title": "CalorieLogResponse"
      },
      "CalorieLogListResponse": {
        "properties": {
          "items": {
            "items": {"$ref": "#/components/schemas/CalorieLogResponse"},
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["items"],
        "title": "CalorieLogListResponse"
      },
      "WorkoutLogCreate": {
        "properties": {
          "record_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Record Date"
          },
          "exercise_id": {
            "type": "string",
            "minLength": 1,
            "title": "Exercise Id"
          },
          "exercise_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "title": "Exercise Name"
          },
          "weight_kg": {
            "anyOf": [
              {"type": "number", "minimum": 0.0, "maximum": 999.0},
              {"type": "null"}
            ],
            "title": "Weight Kg"
          },
          "sets": {
            "type": "array",
            "items": {"type": "integer"},
            "minItems": 1,
            "maxItems": 10,
            "title": "Sets"
          }
        },
        "type": "object",
        "required": ["record_date", "exercise_id", "exercise_name", "sets"],
        "title": "WorkoutLogCreate"
      },
      "WorkoutLogResponse": {
        "properties": {
          "log_id": {"type": "string", "title": "Log Id"},
          "record_date": {"type": "string", "title": "Record Date"},
          "exercise_id": {"type": "string", "title": "Exercise Id"},
          "exercise_name": {"type": "string", "title": "Exercise Name"},
          "weight_kg": {
            "anyOf": [{"type": "number"}, {"type": "null"}],
            "title": "Weight Kg"
          },
          "sets": {
            "type": "array",
            "items": {"type": "integer"},
            "title": "Sets"
          },
          "created_at": {"type": "string", "title": "Created At"},
          "updated_at": {"type": "string", "title": "Updated At"}
        },
        "type": "object",
        "required": ["log_id", "record_date", "exercise_id", "exercise_name", "weight_kg", "sets", "created_at", "updated_at"],
        "title": "WorkoutLogResponse"
      },
      "WorkoutLogListResponse": {
        "properties": {
          "items": {
            "items": {"$ref": "#/components/schemas/WorkoutLogResponse"},
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["items"],
        "title": "WorkoutLogListResponse"
      },
      "WeightRecordCreate": {
        "properties": {
          "record_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Record Date"
          },
          "weight_kg": {
            "type": "number",
            "maximum": 300.0,
            "exclusiveMinimum": 0.0,
            "title": "Weight Kg"
          },
          "body_fat_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Pct"
          }
        },
        "type": "object",
        "required": [
          "record_date",
          "weight_kg"
        ],
        "title": "WeightRecordCreate"
      },
      "WeightRecordListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WeightRecordResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "WeightRecordListResponse"
      },
      "WeightRecordResponse": {
        "properties": {
          "record_date": {
            "type": "string",
            "title": "Record Date"
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg"
          },
          "body_fat_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Pct"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "record_date",
          "weight_kg",
          "body_fat_pct",
          "created_at",
          "updated_at"
        ],
        "title": "WeightRecordResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}